try fixing refactoring 7 day data
Still needs some work, as we're not iterating through protocols properly.
STUN also ends up having TCP and TLS data, which is wrong.
Also realining the trends on the same axis need to be added back in,
and data in Provider needs to be handled properly now that objects have changed.